Updated Tile Placement Rule Enforcement
Now tiles are identified as either being "Travel" or "Six Clearing".  Rule enforcement seems to be done appropriately so far, but at least 1 improvement needs to be made.  When updating which roads lead back to the Borderland, if more than just the tile adjacent to the newly placed tile needs updated, pathfinding and updating needs to be done.  Use Hex IDs for assistance in this matter.

Once that is done, it's all about fine-tuning, and then prepping for going beyond map-building.